refactor: Exercicio 4.2 tem seu proprio arquivo
